Move store folder outside team folder
