+++ src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { BrowserRouter } from "react-router";
 import { Provider } from "react-redux";
-import { store } from "./team/store";
+import { store } from "./store";
 import AppRouter from "./router/AppRouter";
 import "@fontsource/ubuntu/700.css";
 import "@fontsource/ubuntu/400.css";
